refactor osw code
