+++ packages/client/src/observable/internal/base-list/BaseListQuery.ts
@@ -14,14 +14,9 @@
  * limitations under the License.
  */
 
-import type {
-  ObjectSet,
-  ObjectTypeDefinition,
-  Osdk,
-  PageResult,
-} from "@osdk/api";
+import type { ObjectTypeDefinition, Osdk, PageResult } from "@osdk/api";
 import deepEqual from "fast-deep-equal";
-import type { Connectable, Observable, Subscription } from "rxjs";
+import type { Connectable, Observable } from "rxjs";
 import type {
   CommonObserveOptions,
   Status,
@@ -645,93 +640,16 @@ export abstract class BaseListQuery<
   }
 
   //
-  // Shared Websocket Subscription Methods
+  // Websocket Subscription Handler Overrides
   //
 
   /**
-   * Create standard websocket subscription handlers for an ObjectSet.
-   * Subclasses can override individual handlers for custom behavior.
-   *
-   * @param objectSet The ObjectSet to subscribe to
-   * @param sub The parent subscription to add cleanup to
-   * @param methodName The method name for logging purposes
-   */
-  protected createWebsocketSubscription(
-    objectSet: ObjectSet<any>,
-    sub: Subscription,
-    methodName: string = "registerStreamUpdates",
-  ): void {
-    const logger = process.env.NODE_ENV !== "production"
-      ? this.logger?.child({ methodName })
-      : this.logger;
-
-    if (process.env.NODE_ENV !== "production") {
-      logger?.child({ methodName }).info("Subscribing from websocket");
-    }
-
-    try {
-      const websocketSubscription = objectSet.subscribe({
-        onChange: this.onOswChange.bind(this),
-        onError: this.onOswError.bind(this),
-        onOutOfDate: this.onOswOutOfDate.bind(this),
-        onSuccessfulSubscription: this.onOswSuccessfulSubscription.bind(this),
-      });
-
-      sub.add(() => {
-        if (process.env.NODE_ENV !== "production") {
-          logger?.child({ methodName }).info("Unsubscribing from websocket");
-        }
-        websocketSubscription.unsubscribe();
-      });
-    } catch (error) {
-      if (this.logger) {
-        this.logger.child({ methodName })
-          .error("Failed to register stream updates", error);
-      }
-      this.onOswError({ subscriptionClosed: true, error });
-    }
-  }
-
-  /**
-   * Handler called when websocket subscription is successfully established.
-   */
-  protected onOswSuccessfulSubscription(): void {
-    if (process.env.NODE_ENV !== "production") {
-      this.logger?.child({ methodName: "onSuccessfulSubscription" }).debug("");
-    }
-  }
-
-  /**
-   * Handler called when subscribed data becomes out of date.
-   */
-  protected onOswOutOfDate(): void {
-    if (process.env.NODE_ENV !== "production") {
-      this.logger?.child({ methodName: "onOutOfDate" }).debug("");
-    }
-  }
-
-  /**
-   * Handler called when websocket subscription encounters an error.
-   */
-  protected onOswError(errors: {
-    subscriptionClosed: boolean;
-    error: unknown;
-  }): void {
-    if (this.logger) {
-      this.logger?.child({ methodName: "onError" }).error(
-        "subscription errors",
-        errors,
-      );
-    }
-  }
-
-  /**
    * Handler called when an object in the subscribed set is added or updated.
-   * Default implementation stores the object with RDP config if available.
+   * Stores the object with RDP config and handles removal via onOswRemoved.
    *
    * @param update The object update notification
    */
-  protected onOswChange(
+  protected override onOswChange(
     { object, state }: ObjectUpdate<ObjectTypeDefinition, string>,
   ): void {
     const logger = process.env.NODE_ENV !== "production"
